+++ shared-react/src/widgets/index.ts
@@ -5,5 +5,6 @@ export * from "./manager-list";
 export * from "./modal-form";
 export * from "./mount-imperative";
 export * from "./pie-chart";
+export * from "./property-order-table";
 export * from "./property-row";
 export * from "./update-badge";