+++ src/utils/frontmatter/index.ts
@@ -1,5 +1,6 @@
 export * from "./basics";
 export * from "./diff";
+export * from "./ordering";
 export * from "./predicates";
 export * from "./props";
 export * from "./translate-calendar";